moving Category classes to the categories module
